refactor: extract relativeDate to shared util
